demo/demo_02_getdata/get-data/requirements.txt
azure-core==1.14.0b1
azure-storage-blob==12.8.0
azure-storage-file-datalake==12.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
msrest==0.6.21
numpy==1.20.2
oauthlib==3.1.0
pip==21.0.1
pycparser==2.20
requests==2.25.1
requests-oauthlib==1.3.0
setuptools==47.1.0
six==1.15.0
urllib3==1.26.4
